changed (unsigned) to (NSUInteger)
git-svn-id: svn://orca.physics.unc.edu/Orca@11737 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Main Dialogs/CommandCenter/Scripting/ORScriptRunner.h
@@ -84,7 +84,7 @@
 - (void) setFinishCallBack:(id)aTarget selector:(SEL)aSelector;
 - (void) togglePause;
 - (void) singleStep;
-- (unsigned) symbolTableCount;
+- (NSUInteger) symbolTableCount;
 - (id) symbolNameForIndex:(int)i;
 - (id) symbolValueForIndex:(int)i;
 - (long) lastLine;
